pyproject.toml
[tool]
[tool.poetry]
name = "root"
version = "0.0.0"

[tool.poetry.dependencies]
python = "*"
async-timeout = "==3.0.1"
"discord.py" = "*"
dnspython = "==1.16.0"
flask = "*"
openpyxl = "*"
pandas = "*"
pynacl = "==1.3.0"
